Initial commit of new graph engine
New graph engine, written using JavaFX due to its superior graphics
capabilities, such as using double floating point precision used for
coordinates in rendering. This would have been an issue with SWT.

Should be better for our needs than using JGraph, which is very
multipurpose, but not extensible enough for our requirements.

New basis for a UI has been implemented by Sam in JavaFX.

This isn't fully functional, the old code still needs more working
integrating with the new, mostly relating to positioning
nodes. Position nodes was previously delegated to JGraph's automatic
layout, which useful to a degree had the tendency to completely
reorganise the graph when an new node or edge was added. We now need
to provide explicit coordinates, which should be a reasonable guess at
what would look OK. The user shall then be able to rearrange the nodes
to make the graph look better.

Finally, any oddities in the graph code are now our code, which
compared to the size of the codebase of JGraph is comparatively
small. This should therefore be significantly easier to debug.