Handle resizing of canvas (move nodes)
When the user resizes the canvas too small, nodes could go off the
canvas. This patch prevents this by scaling the positions all nodes on
the canvas, the idea is it should preserve the structure of the user's
graph. An alternative would have been to just clamp all nodes in to
the canvas at the edge, however this wouldn't have looked as nice ;)

Includes basic testing application under `test.view` for
GraphCanvasFX.