Ensure that arc rendering assumptions are upheld
As a TODO comment noted, there is an assumption about the height of an
arc and the distance between nodes. This commit adds comments
expanding on the reasoning behind this. This change causes the height
of arcs to scale down when nodes are close, previously, ugly rendering
errors were visible.