Be smart about layout recalculation (fix many TODOs)
During Parham's testing, it was found that performance was poor for
building an Automaton from a regexp with about > 100 states. One issue
was logging, but another is that we always do a full layout
recalculation when we change the automaton. Layout recalculation is
especially expensive in part due to the use of trigonometric functions
to calculate arc angles, sqrt and divide for normalising vectors, etc.

After these changes performance for even 1000 nodes takes ~1 sec.

This change could introduce bugs, testing required on various graph
setups.